% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_signed_area_cd}
\alias{ga_signed_area_cd}
\alias{ga_unsigned_area_cd}
\title{Chamberlain-Duquette spherical area}
\usage{
ga_signed_area_cd(x)

ga_unsigned_area_cd(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a double vector of area values in square meters
}
\description{
These functions use the Chamberlain-Duquette formula, which is suitable for
geographic coordinates on a sphere. Results are in square meters.
}
\examples{
\dontshow{if (requireNamespace("sf", quietly = TRUE) && requireNamespace("geoarrow", quietly = TRUE)) withAutoprint(\{ # examplesIf}
fp <- system.file("shape/nc.shp", package = "sf")
nc <- as.data.frame(read_shapefile(fp))

# square meters, not square degrees
head(as.vector(ga_unsigned_area_cd(nc$geometry)))
head(as.vector(ga_signed_area_cd(nc$geometry)))
\dontshow{\}) # examplesIf}
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/chamberlain_duquette_area/trait.ChamberlainDuquetteArea.html}{ChamberlainDuquetteArea}
}
\seealso{
Other area:
\code{\link[=ga_signed_area]{ga_signed_area()}},
\code{\link[=ga_signed_area_geodesic]{ga_signed_area_geodesic()}}
}
\concept{area}
