% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{signed_area_geodesic}
\alias{signed_area_geodesic}
\alias{unsigned_area_geodesic}
\alias{perimeter_signed_geodesic}
\alias{perimeter_unsigned_geodesic}
\title{Compute the signed and unsigned geodesic area and perimeter of geometries}
\usage{
signed_area_geodesic(x)

unsigned_area_geodesic(x)

perimeter_signed_geodesic(x)

perimeter_unsigned_geodesic(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a double vector of area values in square meters, or perimeter values in meters
}
\description{
These functions use the geodesic formula for computing area and perimeter on
an ellipsoidal model of the earth. Results are in square meters for area and
meters for perimeter.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/geodesic_area/trait.GeodesicArea.html}{GeodesicArea}

\href{https://docs.rs/geo/latest/geo/algorithm/geodesic_area/trait.GeodesicArea.html}{GeodesicArea}

\href{https://docs.rs/geo/latest/geo/algorithm/geodesic_area/trait.GeodesicArea.html}{GeodesicArea}

\href{https://docs.rs/geo/latest/geo/algorithm/geodesic_area/trait.GeodesicArea.html}{GeodesicArea}
}
\seealso{
Other area: 
\code{\link{signed_area}()},
\code{\link{signed_area_cd}()}

Other area: 
\code{\link{signed_area}()},
\code{\link{signed_area_cd}()}

Other area: 
\code{\link{signed_area}()},
\code{\link{signed_area_cd}()}

Other area: 
\code{\link{signed_area}()},
\code{\link{signed_area_cd}()}
}
\concept{area}
