% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bounding_rect}
\alias{bounding_rect}
\title{Compute the axis-aligned bounding rectangle of geometries}
\usage{
bounding_rect(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a GeoArrow rect array
}
\description{
Returns the smallest axis-aligned rectangle that contains each geometry.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/bounding_rect/trait.BoundingRect.html}{BoundingRect}
}
\seealso{
Other boundary: 
\code{\link{minimum_rotated_rect}()}
}
\concept{boundary}
