% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{convex_hull}
\alias{convex_hull}
\title{Compute the convex hull of geometries}
\usage{
convex_hull(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a GeoArrow polygon array
}
\description{
Returns the smallest convex polygon that contains each geometry.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/convex_hull/trait.ConvexHull.html}{ConvexHull}
}
\seealso{
Other boundary:
\code{\link[=bounding_rect]{bounding_rect()}},
\code{\link[=concave_hull]{concave_hull()}},
\code{\link[=extremes]{extremes()}},
\code{\link[=minimum_rotated_rect]{minimum_rotated_rect()}}
}
\concept{boundary}
