% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dist_vincenty_pairwise}
\alias{dist_vincenty_pairwise}
\title{Compute the pairwise Vincenty distance between points}
\usage{
dist_vincenty_pairwise(origin, dest)
}
\arguments{
\item{origin}{a GeoArrow point array of origin points}

\item{dest}{a GeoArrow point array of destination points}
}
\value{
a double vector of distance values in meters
}
\description{
The Vincenty formula computes the geodesic distance between two points on
an ellipsoidal model of the earth. Returns \code{NA} if the algorithm fails to
converge.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/vincenty_distance/trait.VincentyDistance.html}{VincentyDistance}
}
\seealso{
Other distance:
\code{\link[=dist_euclidean_pairwise]{dist_euclidean_pairwise()}},
\code{\link[=dist_frechet_pairwise]{dist_frechet_pairwise()}},
\code{\link[=dist_hausdorff_pairwise]{dist_hausdorff_pairwise()}}
}
\concept{distance}
