% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_buffer}
\alias{ga_buffer}
\title{Buffer geometries by a given distance}
\usage{
ga_buffer(geometry, distance, line_cap, line_join, miter_limit, round_segments)
}
\arguments{
\item{geometry}{a GeoArrow geometry array}

\item{distance}{a numeric vector of buffer distances; length 1 or the same length as \code{geometry}}

\item{line_cap}{one of \code{"round"}, \code{"square"}, or \code{"butt"}}

\item{line_join}{one of \code{"round"}, \code{"miter"}, or \code{"bevel"}}

\item{miter_limit}{the miter limit used when \code{line_join} is \code{"miter"}}

\item{round_segments}{the number of segments used to approximate curves when \code{line_cap} or \code{line_join} is \code{"round"}}
}
\value{
a GeoArrow multipolygon array
}
\description{
Expands each geometry outward by \code{distance} to produce a multipolygon.
Line cap and join styles control the shape of the buffer at endpoints and
corners.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/buffer/trait.Buffer.html}{Buffer}
}
\seealso{
Other misc:
\code{\link[=ga_centroid]{ga_centroid()}},
\code{\link[=ga_chaikin_smoothing]{ga_chaikin_smoothing()}},
\code{\link[=ga_line_segmentize]{ga_line_segmentize()}},
\code{\link[=ga_remove_repeated_points]{ga_remove_repeated_points()}}
}
\concept{misc}
