% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_chaikin_smoothing}
\alias{ga_chaikin_smoothing}
\title{Smooth geometries using the Chaikin algorithm}
\usage{
ga_chaikin_smoothing(geometry, n_iterations)
}
\arguments{
\item{geometry}{a GeoArrow linestring, multilinestring, polygon, or multipolygon array}

\item{n_iterations}{the number of smoothing iterations to apply; must be greater than 0}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Applies Chaikin's corner-cutting algorithm for the given number of iterations
to produce smoother curves. Accepts linestrings, multilinestrings, polygons,
and multipolygons.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/chaikin_smoothing/trait.ChaikinSmoothing.html}{ChaikinSmoothing}
}
\seealso{
Other misc:
\code{\link[=ga_buffer]{ga_buffer()}},
\code{\link[=ga_centroid]{ga_centroid()}},
\code{\link[=ga_line_segmentize]{ga_line_segmentize()}},
\code{\link[=ga_remove_repeated_points]{ga_remove_repeated_points()}}
}
\concept{misc}
