% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_convex_hull}
\alias{ga_convex_hull}
\title{Compute the convex hull of geometries}
\usage{
ga_convex_hull(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a GeoArrow polygon array
}
\description{
Returns the smallest convex polygon that contains each geometry.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/convex_hull/trait.ConvexHull.html}{ConvexHull}
}
\seealso{
Other boundary:
\code{\link[=ga_bounding_rect]{ga_bounding_rect()}},
\code{\link[=ga_concave_hull]{ga_concave_hull()}},
\code{\link[=ga_extremes]{ga_extremes()}},
\code{\link[=ga_minimum_rotated_rect]{ga_minimum_rotated_rect()}}
}
\concept{boundary}
