% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_dist_vincenty_pairwise}
\alias{ga_dist_vincenty_pairwise}
\title{Pairwise Vincenty distance}
\usage{
ga_dist_vincenty_pairwise(origin, dest)
}
\arguments{
\item{origin}{a GeoArrow point array of origin points}

\item{dest}{a GeoArrow point array of destination points}
}
\value{
a double vector of distance values in meters
}
\description{
The Vincenty formula computes the geodesic distance between two points on
an ellipsoidal model of the earth. Returns \code{NA} if the algorithm fails to
converge.
}
\examples{
\dontshow{if (requireNamespace("sf", quietly = TRUE) && requireNamespace("geoarrow", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Raleigh to Charlotte, and Raleigh to Wilmington
origin <- ga_xy(c(-78.6382, -78.6382), c(35.7796, 35.7796))
dest <- ga_xy(c(-80.8431, -77.9447), c(35.2271, 34.2257))

as.vector(ga_dist_vincenty_pairwise(origin, dest))
\dontshow{\}) # examplesIf}
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/vincenty_distance/trait.VincentyDistance.html}{VincentyDistance}
}
\seealso{
Other distance:
\code{\link[=ga_dist_euclidean_pairwise]{ga_dist_euclidean_pairwise()}},
\code{\link[=ga_dist_frechet_pairwise]{ga_dist_frechet_pairwise()}},
\code{\link[=ga_dist_hausdorff_pairwise]{ga_dist_hausdorff_pairwise()}}
}
\concept{distance}
