% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_line_segmentize}
\alias{ga_line_segmentize}
\alias{ga_line_segmentize_haversine}
\title{Split linestrings into a given number of equal-length segments}
\usage{
ga_line_segmentize(geometry, segment_count)

ga_line_segmentize_haversine(geometry, segment_count)
}
\arguments{
\item{geometry}{a GeoArrow linestring array}

\item{segment_count}{the number of segments to split each linestring into; must be greater than 0}
}
\value{
a GeoArrow multilinestring array
}
\description{
Divides each linestring into \code{segment_count} segments of equal length,
returning a multilinestring. Returns \code{NA} if segmentation fails.
\code{ga_line_segmentize_haversine()} uses the Haversine formula for geographic
coordinates; \code{ga_line_segmentize()} uses planar Euclidean distance.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/linestring_segment/trait.LineStringSegmentize.html}{LineStringSegmentize}

\href{https://docs.rs/geo/latest/geo/algorithm/linestring_segment/trait.LineStringSegmentizeHaversine.html}{LineStringSegmentizeHaversine}
}
\seealso{
Other misc:
\code{\link[=ga_buffer]{ga_buffer()}},
\code{\link[=ga_centroid]{ga_centroid()}},
\code{\link[=ga_chaikin_smoothing]{ga_chaikin_smoothing()}},
\code{\link[=ga_remove_repeated_points]{ga_remove_repeated_points()}}
}
\concept{misc}
