% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_rotate_around_center}
\alias{ga_rotate_around_center}
\title{Rotate around the bounding box center}
\usage{
ga_rotate_around_center(geometry, degrees)
}
\arguments{
\item{geometry}{a GeoArrow point, linestring, multilinestring, polygon, or multipolygon array}

\item{degrees}{angle of rotation; length 1 or the same length as \code{geometry}}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Rotates each geometry counter-clockwise by \code{degrees} about the center of its
own axis-aligned bounding rectangle. The geometry type of the output matches
the geometry type of the input.
}
\examples{
\dontshow{if (requireNamespace("sf", quietly = TRUE) && requireNamespace("geoarrow", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# the same triangle, turned about its bounding box center instead
tri <- geoarrow::as_geoarrow_array(sf::st_sfc(
  sf::st_polygon(list(rbind(c(0, 0), c(4, 0), c(0, 3), c(0, 0))))
))

geoarrow::as_geoarrow_vctr(ga_rotate_around_center(tri, 90))
\dontshow{\}) # examplesIf}
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/rotate/trait.Rotate.html}{Rotate}
}
\seealso{
Other affine:
\code{\link[=ga_affine_transform]{ga_affine_transform()}},
\code{\link[=ga_rotate_around_centroid]{ga_rotate_around_centroid()}},
\code{\link[=ga_scale_xy]{ga_scale_xy()}},
\code{\link[=ga_skew]{ga_skew()}},
\code{\link[=ga_skew_xy]{ga_skew_xy()}},
\code{\link[=ga_translate]{ga_translate()}}
}
\concept{affine}
