% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_scale_xy}
\alias{ga_scale_xy}
\title{Scale geometries independently in x and y about their centroid}
\usage{
ga_scale_xy(geometry, x_factor, y_factor)
}
\arguments{
\item{geometry}{a GeoArrow point, linestring, multilinestring, polygon, or multipolygon array}

\item{x_factor}{scaling factor along the x axis; length 1 or the same length as \code{geometry}}

\item{y_factor}{scaling factor along the y axis; length 1 or the same length as \code{geometry}}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Scales each geometry by \code{x_factor} along the x axis and \code{y_factor} along the
y axis, about the geometry's own centroid. The geometry type of the output
matches the geometry type of the input.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/scale/trait.Scale.html}{Scale}
}
\seealso{
Other affine:
\code{\link[=ga_affine_transform]{ga_affine_transform()}},
\code{\link[=ga_rotate_around_center]{ga_rotate_around_center()}},
\code{\link[=ga_rotate_around_centroid]{ga_rotate_around_centroid()}},
\code{\link[=ga_skew]{ga_skew()}},
\code{\link[=ga_skew_xy]{ga_skew_xy()}},
\code{\link[=ga_translate]{ga_translate()}}
}
\concept{affine}
