% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_skew_xy}
\alias{ga_skew_xy}
\title{Skew geometries independently in x and y about their centroid}
\usage{
ga_skew_xy(geometry, degrees_x, degrees_y)
}
\arguments{
\item{geometry}{a GeoArrow point, linestring, multilinestring, polygon, or multipolygon array}

\item{degrees_x}{shear angle along the x dimension; length 1 or the same length as \code{geometry}}

\item{degrees_y}{shear angle along the y dimension; length 1 or the same length as \code{geometry}}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Shears each geometry by \code{degrees_x} along the x dimension and \code{degrees_y}
along the y dimension, about the geometry's own centroid. The geometry type
of the output matches the geometry type of the input.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/skew/trait.Skew.html}{Skew}
}
\seealso{
Other affine:
\code{\link[=ga_affine_transform]{ga_affine_transform()}},
\code{\link[=ga_rotate_around_center]{ga_rotate_around_center()}},
\code{\link[=ga_rotate_around_centroid]{ga_rotate_around_centroid()}},
\code{\link[=ga_scale_xy]{ga_scale_xy()}},
\code{\link[=ga_skew]{ga_skew()}},
\code{\link[=ga_translate]{ga_translate()}}
}
\concept{affine}
