% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_translate}
\alias{ga_translate}
\title{Translate geometries along the x and y axes}
\usage{
ga_translate(geometry, x_offset, y_offset)
}
\arguments{
\item{geometry}{a GeoArrow point, linestring, multilinestring, polygon, or multipolygon array}

\item{x_offset}{distance to shift along the x axis; length 1 or the same length as \code{geometry}}

\item{y_offset}{distance to shift along the y axis; length 1 or the same length as \code{geometry}}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Shifts every coordinate of each geometry by \code{x_offset} and \code{y_offset}. The
geometry type of the output matches the geometry type of the input.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/translate/trait.Translate.html}{Translate}
}
\seealso{
Other affine:
\code{\link[=ga_affine_transform]{ga_affine_transform()}},
\code{\link[=ga_rotate_around_center]{ga_rotate_around_center()}},
\code{\link[=ga_rotate_around_centroid]{ga_rotate_around_centroid()}},
\code{\link[=ga_scale_xy]{ga_scale_xy()}},
\code{\link[=ga_skew]{ga_skew()}},
\code{\link[=ga_skew_xy]{ga_skew_xy()}}
}
\concept{affine}
