% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{minimum_rotated_rect}
\alias{minimum_rotated_rect}
\title{Compute the minimum rotated bounding rectangle of geometries}
\usage{
minimum_rotated_rect(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a GeoArrow polygon array
}
\description{
Returns the smallest rectangle of arbitrary rotation that contains each
geometry, as a polygon array.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/minimum_rotated_rect/trait.MinimumRotatedRect.html}{MinimumRotatedRect}
}
\seealso{
Other boundary: 
\code{\link{bounding_rect}()}
}
\concept{boundary}
