% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{points_along_line}
\alias{points_along_line}
\title{Generate points at regular intervals along the line between two points}
\usage{
points_along_line(start, end, max_distance, include_ends, metric)
}
\arguments{
\item{start}{a GeoArrow point array of start points}

\item{end}{a GeoArrow point array of end points}

\item{max_distance}{the maximum spacing between generated points; length 1 or the same length as \code{start}}

\item{include_ends}{whether to include the start and end points in the output}

\item{metric}{one of \code{"euclidean"}, \code{"haversine"}, \code{"geodesic"}, or \code{"rhumb"}}
}
\value{
a GeoArrow multipoint array
}
\description{
Returns a multipoint array where each element contains all points spaced
at most \code{max_distance} apart along the path from \code{start} to \code{end}.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.InterpolatePoint.html}{InterpolatePoint}
}
\seealso{
Other interpolate:
\code{\link[=interpolate_point]{interpolate_point()}},
\code{\link[=point_at_distance_between]{point_at_distance_between()}}
}
\concept{interpolate}
