% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{to_degrees}
\alias{to_degrees}
\title{Convert coordinates from radians to degrees}
\usage{
to_degrees(geometry)
}
\arguments{
\item{geometry}{a GeoArrow geometry array}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Multiplies every x and y coordinate by \code{180 / pi}. The geometry type of the
output matches the geometry type of the input.
}
\details{
Only the x and y coordinates are converted. Z and M values are dropped,
since the conversion goes through a two dimensional representation.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/convert_angle_unit/trait.ToDegrees.html}{ToDegrees}
}
\seealso{
Other convert:
\code{\link[=to_radians]{to_radians()}}
}
\concept{convert}
