% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{translate}
\alias{translate}
\title{Translate geometries along the x and y axes}
\usage{
translate(geometry, x_offset, y_offset)
}
\arguments{
\item{geometry}{a GeoArrow point, linestring, multilinestring, polygon, or multipolygon array}

\item{x_offset}{distance to shift along the x axis; length 1 or the same length as \code{geometry}}

\item{y_offset}{distance to shift along the y axis; length 1 or the same length as \code{geometry}}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Shifts every coordinate of each geometry by \code{x_offset} and \code{y_offset}. The
geometry type of the output matches the geometry type of the input.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/translate/trait.Translate.html}{Translate}
}
\seealso{
Other affine:
\code{\link[=affine_transform]{affine_transform()}},
\code{\link[=rotate_around_center]{rotate_around_center()}},
\code{\link[=rotate_around_centroid]{rotate_around_centroid()}},
\code{\link[=scale_xy]{scale_xy()}},
\code{\link[=skew]{skew()}},
\code{\link[=skew_xy]{skew_xy()}}
}
\concept{affine}
