% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_signed_area}
\alias{ga_signed_area}
\alias{ga_unsigned_area}
\title{Signed and unsigned planar area}
\usage{
ga_signed_area(x)

ga_unsigned_area(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a double vector of area values in the units of the coordinate system
}
\description{
\code{ga_signed_area()} returns positive values for counter-clockwise winding and
negative values for clockwise winding. \code{ga_unsigned_area()} always returns a
non-negative value.
}
\examples{
\dontshow{if (requireNamespace("sf", quietly = TRUE) && requireNamespace("geoarrow", quietly = TRUE)) withAutoprint(\{ # examplesIf}
fp <- system.file("shape/nc.shp", package = "sf")
nc <- as.data.frame(read_shapefile(fp))

# nc rings wind clockwise, so the signed area is negative
head(as.vector(ga_signed_area(nc$geometry)))
head(as.vector(ga_unsigned_area(nc$geometry)))
\dontshow{\}) # examplesIf}
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/area/trait.Area.html}{Area}
}
\seealso{
Other area:
\code{\link[=ga_signed_area_cd]{ga_signed_area_cd()}},
\code{\link[=ga_signed_area_geodesic]{ga_signed_area_geodesic()}}
}
\concept{area}
