% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{signed_area}
\alias{signed_area}
\alias{unsigned_area}
\title{Compute the signed and unsigned planar area of geometries}
\usage{
signed_area(x)

unsigned_area(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a double vector of area values in the units of the coordinate system
}
\description{
\code{signed_area()} returns positive values for counter-clockwise winding and
negative values for clockwise winding. \code{unsigned_area()} always returns a
non-negative value.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/area/trait.Area.html}{Area}

\href{https://docs.rs/geo/latest/geo/algorithm/area/trait.Area.html}{Area}
}
\seealso{
Other area: 
\code{\link{signed_area_cd}()},
\code{\link{signed_area_geodesic}()}

Other area: 
\code{\link{signed_area_cd}()},
\code{\link{signed_area_geodesic}()}
}
\concept{area}
