% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_signed_area_cd}
\alias{ga_signed_area_cd}
\alias{ga_unsigned_area_cd}
\title{Compute the signed and unsigned area using the Chamberlain-Duquette algorithm}
\usage{
ga_signed_area_cd(x)

ga_unsigned_area_cd(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a double vector of area values in square meters
}
\description{
These functions use the Chamberlain-Duquette formula, which is suitable for
geographic coordinates on a sphere. Results are in square meters.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/chamberlain_duquette_area/trait.ChamberlainDuquetteArea.html}{ChamberlainDuquetteArea}

\href{https://docs.rs/geo/latest/geo/algorithm/chamberlain_duquette_area/trait.ChamberlainDuquetteArea.html}{ChamberlainDuquetteArea}
}
\seealso{
Other area:
\code{\link[=ga_signed_area]{ga_signed_area()}},
\code{\link[=ga_signed_area_geodesic]{ga_signed_area_geodesic()}}
}
\concept{area}
