% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bearing_euclidean}
\alias{bearing_euclidean}
\alias{bearing_haversine}
\alias{bearing_geodesic}
\alias{bearing_rhumb}
\title{Compute the bearing between pairs of points}
\usage{
bearing_euclidean(origin, dest)

bearing_haversine(origin, dest)

bearing_geodesic(origin, dest)

bearing_rhumb(origin, dest)
}
\arguments{
\item{origin}{a GeoArrow point array of origin points}

\item{dest}{a GeoArrow point array of destination points}
}
\value{
a double vector of bearing values in degrees
}
\description{
Calculate the bearing in degrees from \code{origin} to \code{dest} for each pair of
points. Bearing is measured clockwise from north (0 degrees) to 360 degrees.
These functions differ in the metric space used for the calculation.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.Bearing.html}{Bearing},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/struct.Euclidean.html}{Euclidean}

\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.Bearing.html}{Bearing},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/constant.Haversine.html}{Haversine}

\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.Bearing.html}{Bearing},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/static.Geodesic.html}{Geodesic}

\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.Bearing.html}{Bearing},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/struct.Rhumb.html}{Rhumb}
}
\concept{bearing}
