% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{centroid}
\alias{centroid}
\title{Compute the centroid of geometries}
\usage{
centroid(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a GeoArrow point array
}
\description{
Returns the centroid point of each geometry. Returns \code{NA} for empty geometries.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/centroid/trait.Centroid.html}{Centroid}
}
\seealso{
Other misc:
\code{\link[=buffer]{buffer()}},
\code{\link[=chaikin_smoothing]{chaikin_smoothing()}},
\code{\link[=line_segmentize]{line_segmentize()}},
\code{\link[=remove_repeated_points]{remove_repeated_points()}}
}
\concept{misc}
