% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{chaikin_smoothing}
\alias{chaikin_smoothing}
\title{Smooth geometries using the Chaikin algorithm}
\usage{
chaikin_smoothing(geometry, n_iterations)
}
\arguments{
\item{geometry}{a GeoArrow linestring, multilinestring, polygon, or multipolygon array}

\item{n_iterations}{the number of smoothing iterations to apply; must be greater than 0}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Applies Chaikin's corner-cutting algorithm for the given number of iterations
to produce smoother curves. Accepts linestrings, multilinestrings, polygons,
and multipolygons.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/chaikin_smoothing/trait.ChaikinSmoothing.html}{ChaikinSmoothing}
}
\seealso{
Other misc: 
\code{\link{buffer}()},
\code{\link{centroid}()},
\code{\link{line_segmentize}()},
\code{\link{remove_repeated_points}()}
}
\concept{misc}
