% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dist_frechet_pairwise}
\alias{dist_frechet_pairwise}
\title{Compute the pairwise Frechet distance between linestrings}
\usage{
dist_frechet_pairwise(origin, dest)
}
\arguments{
\item{origin}{a GeoArrow linestring array}

\item{dest}{a GeoArrow linestring array}
}
\value{
a double vector of Frechet distance values
}
\description{
The Frechet distance measures the similarity between two curves by
considering the location and ordering of points along each curve.
Uses the Euclidean metric.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.FrechetDistance.html}{FrechetDistance}
}
\seealso{
Other distance:
\code{\link[=dist_euclidean_pairwise]{dist_euclidean_pairwise()}},
\code{\link[=dist_hausdorff_pairwise]{dist_hausdorff_pairwise()}},
\code{\link[=dist_vincenty_pairwise]{dist_vincenty_pairwise()}}
}
\concept{distance}
