% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dist_hausdorff_pairwise}
\alias{dist_hausdorff_pairwise}
\title{Compute the pairwise Hausdorff distance between geometries}
\usage{
dist_hausdorff_pairwise(origin, dest)
}
\arguments{
\item{origin}{a GeoArrow geometry array}

\item{dest}{a GeoArrow geometry array}
}
\value{
a double vector of Hausdorff distance values
}
\description{
The Hausdorff distance measures how far two geometries are from each other
by taking the maximum of all minimum distances between points on the two shapes.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/hausdorff_distance/trait.HausdorffDistance.html}{HausdorffDistance}
}
\seealso{
Other distance: 
\code{\link{dist_euclidean_pairwise}()},
\code{\link{dist_frechet_pairwise}()},
\code{\link{dist_vincenty_pairwise}()}
}
\concept{distance}
