% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_dist_euclidean_pairwise}
\alias{ga_dist_euclidean_pairwise}
\alias{ga_dist_haversine_pairwise}
\alias{ga_dist_geodesic_pairwise}
\alias{ga_dist_rhumb_pairwise}
\title{Compute pairwise distances between points}
\usage{
ga_dist_euclidean_pairwise(origin, dest)

ga_dist_haversine_pairwise(origin, dest)

ga_dist_geodesic_pairwise(origin, dest)

ga_dist_rhumb_pairwise(origin, dest)
}
\arguments{
\item{origin}{a GeoArrow point array of origin points}

\item{dest}{a GeoArrow point array of destination points}
}
\value{
a double vector of distance values
}
\description{
Calculate the distance between each pair of points in \code{origin} and \code{dest}.
These functions differ in the metric space used for the calculation.
}
\examples{
\dontshow{if (requireNamespace("sf", quietly = TRUE) && requireNamespace("geoarrow", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Raleigh to Charlotte, and Raleigh to Wilmington
origin <- ga_xy(c(-78.6382, -78.6382), c(35.7796, 35.7796))
dest <- ga_xy(c(-80.8431, -77.9447), c(35.2271, 34.2257))

# degrees, then meters
as.vector(ga_dist_euclidean_pairwise(origin, dest))
as.vector(ga_dist_geodesic_pairwise(origin, dest))
\dontshow{\}) # examplesIf}
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.Distance.html}{Distance},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/struct.Euclidean.html}{Euclidean},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/constant.Haversine.html}{Haversine},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/static.Geodesic.html}{Geodesic},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/struct.Rhumb.html}{Rhumb}
}
\seealso{
Other distance:
\code{\link[=ga_dist_frechet_pairwise]{ga_dist_frechet_pairwise()}},
\code{\link[=ga_dist_hausdorff_pairwise]{ga_dist_hausdorff_pairwise()}},
\code{\link[=ga_dist_vincenty_pairwise]{ga_dist_vincenty_pairwise()}}
}
\concept{distance}
