% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dist_euclidean_pairwise}
\alias{dist_euclidean_pairwise}
\alias{dist_haversine_pairwise}
\alias{dist_geodesic_pairwise}
\alias{dist_rhumb_pairwise}
\title{Compute pairwise distances between points}
\usage{
dist_euclidean_pairwise(origin, dest)

dist_haversine_pairwise(origin, dest)

dist_geodesic_pairwise(origin, dest)

dist_rhumb_pairwise(origin, dest)
}
\arguments{
\item{origin}{a GeoArrow point array of origin points}

\item{dest}{a GeoArrow point array of destination points}
}
\value{
a double vector of distance values
}
\description{
Calculate the distance between each pair of points in \code{origin} and \code{dest}.
These functions differ in the metric space used for the calculation.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.Distance.html}{Distance},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/struct.Euclidean.html}{Euclidean}

\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.Distance.html}{Distance},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/constant.Haversine.html}{Haversine}

\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.Distance.html}{Distance},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/static.Geodesic.html}{Geodesic}

\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.Distance.html}{Distance},
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/metric_spaces/struct.Rhumb.html}{Rhumb}
}
\seealso{
Other distance:
\code{\link[=dist_frechet_pairwise]{dist_frechet_pairwise()}},
\code{\link[=dist_hausdorff_pairwise]{dist_hausdorff_pairwise()}},
\code{\link[=dist_vincenty_pairwise]{dist_vincenty_pairwise()}}
}
\concept{distance}
