% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_bounding_rect}
\alias{ga_bounding_rect}
\title{Axis-aligned bounding rectangle}
\usage{
ga_bounding_rect(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a GeoArrow rect array
}
\description{
Returns the smallest axis-aligned rectangle that contains each geometry.
}
\examples{
\dontshow{if (requireNamespace("sf", quietly = TRUE) && requireNamespace("geoarrow", quietly = TRUE)) withAutoprint(\{ # examplesIf}
fp <- system.file("shape/nc.shp", package = "sf")
nc <- as.data.frame(read_shapefile(fp))

head(geoarrow::as_geoarrow_vctr(ga_bounding_rect(nc$geometry)), 3)
\dontshow{\}) # examplesIf}
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/bounding_rect/trait.BoundingRect.html}{BoundingRect}
}
\seealso{
Other boundary:
\code{\link[=ga_concave_hull]{ga_concave_hull()}},
\code{\link[=ga_convex_hull]{ga_convex_hull()}},
\code{\link[=ga_extremes]{ga_extremes()}},
\code{\link[=ga_minimum_rotated_rect]{ga_minimum_rotated_rect()}}
}
\concept{boundary}
