% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_centroid}
\alias{ga_centroid}
\title{Compute the centroid of geometries}
\usage{
ga_centroid(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a GeoArrow point array
}
\description{
Returns the centroid point of each geometry. Returns \code{NA} for empty geometries.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/centroid/trait.Centroid.html}{Centroid}
}
\seealso{
Other misc:
\code{\link[=ga_buffer]{ga_buffer()}},
\code{\link[=ga_chaikin_smoothing]{ga_chaikin_smoothing()}},
\code{\link[=ga_line_segmentize]{ga_line_segmentize()}},
\code{\link[=ga_remove_repeated_points]{ga_remove_repeated_points()}}
}
\concept{misc}
