% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_dist_hausdorff_pairwise}
\alias{ga_dist_hausdorff_pairwise}
\title{Compute the pairwise Hausdorff distance between geometries}
\usage{
ga_dist_hausdorff_pairwise(origin, dest)
}
\arguments{
\item{origin}{a GeoArrow geometry array}

\item{dest}{a GeoArrow geometry array}
}
\value{
a double vector of Hausdorff distance values
}
\description{
The Hausdorff distance measures how far two geometries are from each other
by taking the maximum of all minimum distances between points on the two shapes.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/hausdorff_distance/trait.HausdorffDistance.html}{HausdorffDistance}
}
\seealso{
Other distance:
\code{\link[=ga_dist_euclidean_pairwise]{ga_dist_euclidean_pairwise()}},
\code{\link[=ga_dist_frechet_pairwise]{ga_dist_frechet_pairwise()}},
\code{\link[=ga_dist_vincenty_pairwise]{ga_dist_vincenty_pairwise()}}
}
\concept{distance}
