% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_dist_hausdorff_pairwise}
\alias{ga_dist_hausdorff_pairwise}
\title{Pairwise Hausdorff distance}
\usage{
ga_dist_hausdorff_pairwise(origin, dest)
}
\arguments{
\item{origin}{a GeoArrow geometry array}

\item{dest}{a GeoArrow geometry array; length 1 or the same length as \code{origin}}
}
\value{
a double vector of Hausdorff distance values
}
\description{
The Hausdorff distance measures how far two geometries are from each other
by taking the maximum of all minimum distances between points on the two shapes.
}
\examples{
\dontshow{if (requireNamespace("sf", quietly = TRUE) && requireNamespace("geoarrow", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# how far apart are neighbouring counties at their worst?
fp <- system.file("shape/nc.shp", package = "sf")
nc <- as.data.frame(read_shapefile(fp))

as.vector(ga_dist_hausdorff_pairwise(nc$geometry[1:5], nc$geometry[2:6]))
\dontshow{\}) # examplesIf}
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/hausdorff_distance/trait.HausdorffDistance.html}{HausdorffDistance}
}
\seealso{
Other distance:
\code{\link[=ga_dist_euclidean_pairwise]{ga_dist_euclidean_pairwise()}},
\code{\link[=ga_dist_frechet_pairwise]{ga_dist_frechet_pairwise()}},
\code{\link[=ga_dist_vincenty_pairwise]{ga_dist_vincenty_pairwise()}}
}
\concept{distance}
