% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_interpolate_point}
\alias{ga_interpolate_point}
\title{Interpolate a point along a linestring}
\usage{
ga_interpolate_point(line, value, metric, measure, from)
}
\arguments{
\item{line}{a GeoArrow linestring array}

\item{value}{a numeric vector of ratio or distance values; length 1 or the same length as \code{line}}

\item{metric}{one of \code{"euclidean"}, \code{"haversine"}, \code{"geodesic"}, or \code{"rhumb"}}

\item{measure}{one of \code{"ratio"} or \code{"distance"}}

\item{from}{one of \code{"start"} or \code{"end"}}
}
\value{
a GeoArrow point array
}
\description{
Returns the point at a given ratio or distance along each linestring,
measured from either the start or end. The metric determines how distance
is calculated.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/line_measures/trait.InterpolateLine.html}{InterpolateLine}
}
\seealso{
Other interpolate:
\code{\link[=ga_point_at_distance_between]{ga_point_at_distance_between()}},
\code{\link[=ga_points_along_line]{ga_points_along_line()}}
}
\concept{interpolate}
