% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_minimum_rotated_rect}
\alias{ga_minimum_rotated_rect}
\title{Compute the minimum rotated bounding rectangle of geometries}
\usage{
ga_minimum_rotated_rect(x)
}
\arguments{
\item{x}{a GeoArrow geometry array}
}
\value{
a GeoArrow polygon array
}
\description{
Returns the smallest rectangle of arbitrary rotation that contains each
geometry, as a polygon array.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/minimum_rotated_rect/trait.MinimumRotatedRect.html}{MinimumRotatedRect}
}
\seealso{
Other boundary:
\code{\link[=ga_bounding_rect]{ga_bounding_rect()}},
\code{\link[=ga_concave_hull]{ga_concave_hull()}},
\code{\link[=ga_convex_hull]{ga_convex_hull()}},
\code{\link[=ga_extremes]{ga_extremes()}}
}
\concept{boundary}
