% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_remove_repeated_points}
\alias{ga_remove_repeated_points}
\title{Remove repeated consecutive points from geometries}
\usage{
ga_remove_repeated_points(geometry)
}
\arguments{
\item{geometry}{a GeoArrow linestring, multilinestring, polygon, or multipolygon array}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Removes consecutive duplicate coordinates from each geometry. Accepts
linestrings, multilinestrings, polygons, and multipolygons.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/remove_repeated_points/trait.RemoveRepeatedPoints.html}{RemoveRepeatedPoints}
}
\seealso{
Other misc:
\code{\link[=ga_buffer]{ga_buffer()}},
\code{\link[=ga_centroid]{ga_centroid()}},
\code{\link[=ga_chaikin_smoothing]{ga_chaikin_smoothing()}},
\code{\link[=ga_line_segmentize]{ga_line_segmentize()}}
}
\concept{misc}
