% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_rotate_around_center}
\alias{ga_rotate_around_center}
\title{Rotate geometries around the center of their bounding box}
\usage{
ga_rotate_around_center(geometry, degrees)
}
\arguments{
\item{geometry}{a GeoArrow point, linestring, multilinestring, polygon, or multipolygon array}

\item{degrees}{angle of rotation; length 1 or the same length as \code{geometry}}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Rotates each geometry counter-clockwise by \code{degrees} about the center of its
own axis-aligned bounding rectangle. The geometry type of the output matches
the geometry type of the input.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/rotate/trait.Rotate.html}{Rotate}
}
\seealso{
Other affine:
\code{\link[=ga_affine_transform]{ga_affine_transform()}},
\code{\link[=ga_rotate_around_centroid]{ga_rotate_around_centroid()}},
\code{\link[=ga_scale_xy]{ga_scale_xy()}},
\code{\link[=ga_skew]{ga_skew()}},
\code{\link[=ga_skew_xy]{ga_skew_xy()}},
\code{\link[=ga_translate]{ga_translate()}}
}
\concept{affine}
