% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_simplify}
\alias{ga_simplify}
\title{Simplify geometries using the Ramer-Douglas-Peucker algorithm}
\usage{
ga_simplify(geometry, epsilon)
}
\arguments{
\item{geometry}{a GeoArrow linestring, multilinestring, polygon, or multipolygon array}

\item{epsilon}{the simplification tolerance; length 1 or the same length as \code{geometry}}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Reduces the number of points in each geometry by removing vertices that
deviate less than \code{epsilon} from the simplified path. Accepts linestrings,
multilinestrings, polygons, and multipolygons.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/simplify/trait.Simplify.html}{Simplify}
}
\seealso{
Other simplify:
\code{\link[=ga_simplify_idx]{ga_simplify_idx()}},
\code{\link[=ga_simplify_vw]{ga_simplify_vw()}}
}
\concept{simplify}
