% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ga_to_radians}
\alias{ga_to_radians}
\title{Convert coordinates from degrees to radians}
\usage{
ga_to_radians(geometry)
}
\arguments{
\item{geometry}{a GeoArrow geometry array}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Multiplies every x and y coordinate by \code{pi / 180}. The geometry type of the
output matches the geometry type of the input.
}
\details{
Only the x and y coordinates are converted. Z and M values are dropped,
since the conversion goes through a two dimensional representation.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/convert_angle_unit/trait.ToRadians.html}{ToRadians}
}
\seealso{
Other convert:
\code{\link[=ga_to_degrees]{ga_to_degrees()}}
}
\concept{convert}
