% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{line_segmentize}
\alias{line_segmentize}
\alias{line_segmentize_haversine}
\title{Split linestrings into a given number of equal-length segments}
\usage{
line_segmentize(geometry, segment_count)

line_segmentize_haversine(geometry, segment_count)
}
\arguments{
\item{geometry}{a GeoArrow linestring array}

\item{segment_count}{the number of segments to split each linestring into; must be greater than 0}
}
\value{
a GeoArrow multilinestring array
}
\description{
Divides each linestring into \code{segment_count} segments of equal length,
returning a multilinestring. Returns \code{NA} if segmentation fails.
\code{line_segmentize_haversine()} uses the Haversine formula for geographic
coordinates; \code{line_segmentize()} uses planar Euclidean distance.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/linestring_segment/trait.LineStringSegmentize.html}{LineStringSegmentize}

\href{https://docs.rs/geo/latest/geo/algorithm/linestring_segment/trait.LineStringSegmentizeHaversine.html}{LineStringSegmentizeHaversine}
}
\seealso{
Other misc: 
\code{\link{buffer}()},
\code{\link{centroid}()},
\code{\link{chaikin_smoothing}()},
\code{\link{remove_repeated_points}()}

Other misc: 
\code{\link{buffer}()},
\code{\link{centroid}()},
\code{\link{chaikin_smoothing}()},
\code{\link{remove_repeated_points}()}
}
\concept{misc}
