% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{remove_repeated_points}
\alias{remove_repeated_points}
\title{Remove repeated consecutive points from geometries}
\usage{
remove_repeated_points(geometry)
}
\arguments{
\item{geometry}{a GeoArrow linestring, multilinestring, polygon, or multipolygon array}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Removes consecutive duplicate coordinates from each geometry. Accepts
linestrings, multilinestrings, polygons, and multipolygons.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/remove_repeated_points/trait.RemoveRepeatedPoints.html}{RemoveRepeatedPoints}
}
\seealso{
Other misc:
\code{\link[=buffer]{buffer()}},
\code{\link[=centroid]{centroid()}},
\code{\link[=chaikin_smoothing]{chaikin_smoothing()}},
\code{\link[=line_segmentize]{line_segmentize()}}
}
\concept{misc}
