% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rotate_around_center}
\alias{rotate_around_center}
\title{Rotate geometries around the center of their bounding box}
\usage{
rotate_around_center(geometry, degrees)
}
\arguments{
\item{geometry}{a GeoArrow point, linestring, multilinestring, polygon, or multipolygon array}

\item{degrees}{angle of rotation; length 1 or the same length as \code{geometry}}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Rotates each geometry counter-clockwise by \code{degrees} about the center of its
own axis-aligned bounding rectangle. The geometry type of the output matches
the geometry type of the input.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/rotate/trait.Rotate.html}{Rotate}
}
\seealso{
Other affine:
\code{\link[=affine_transform]{affine_transform()}},
\code{\link[=rotate_around_centroid]{rotate_around_centroid()}},
\code{\link[=scale_xy]{scale_xy()}},
\code{\link[=skew]{skew()}},
\code{\link[=skew_xy]{skew_xy()}},
\code{\link[=translate]{translate()}}
}
\concept{affine}
