% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{simplify_vw}
\alias{simplify_vw}
\alias{simplify_vw_preserve}
\title{Simplify geometries using the Visvalingam-Whyatt algorithm}
\usage{
simplify_vw(geometry, epsilon)

simplify_vw_preserve(geometry, epsilon)
}
\arguments{
\item{geometry}{a GeoArrow linestring, multilinestring, polygon, or multipolygon array}

\item{epsilon}{the simplification tolerance; length 1 or the same length as \code{geometry}}
}
\value{
a GeoArrow array of the same geometry type as the input

a GeoArrow array of the same geometry type as the input
}
\description{
Reduces the number of points in each geometry by removing vertices whose
effective area is below \code{epsilon}. Accepts linestrings, multilinestrings,
polygons, and multipolygons.

Like \code{simplify_vw()} but preserves topology by preventing self-intersections
during simplification. Accepts linestrings, multilinestrings, polygons, and
multipolygons.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/simplify_vw/trait.SimplifyVw.html}{SimplifyVw}

\href{https://docs.rs/geo/latest/geo/algorithm/simplify_vw/trait.SimplifyVwPreserve.html}{SimplifyVwPreserve}
}
\seealso{
Other simplify: 
\code{\link{simplify}()}

Other simplify: 
\code{\link{simplify}()}
}
\concept{simplify}
