% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{skew}
\alias{skew}
\title{Skew geometries uniformly about their centroid}
\usage{
skew(geometry, degrees)
}
\arguments{
\item{geometry}{a GeoArrow point, linestring, multilinestring, polygon, or multipolygon array}

\item{degrees}{the shear angle; length 1 or the same length as \code{geometry}}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Shears each geometry by \code{degrees} along both the x and y dimensions, about
the geometry's own centroid. The geometry type of the output matches the
geometry type of the input.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/skew/trait.Skew.html}{Skew}
}
\seealso{
Other affine:
\code{\link[=rotate_around_center]{rotate_around_center()}},
\code{\link[=rotate_around_centroid]{rotate_around_centroid()}},
\code{\link[=scale_xy]{scale_xy()}},
\code{\link[=skew_xy]{skew_xy()}},
\code{\link[=translate]{translate()}}
}
\concept{affine}
