% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{skew_xy}
\alias{skew_xy}
\title{Skew geometries independently in x and y about their centroid}
\usage{
skew_xy(geometry, degrees_x, degrees_y)
}
\arguments{
\item{geometry}{a GeoArrow point, linestring, multilinestring, polygon, or multipolygon array}

\item{degrees_x}{shear angle along the x dimension; length 1 or the same length as \code{geometry}}

\item{degrees_y}{shear angle along the y dimension; length 1 or the same length as \code{geometry}}
}
\value{
a GeoArrow array of the same geometry type as the input
}
\description{
Shears each geometry by \code{degrees_x} along the x dimension and \code{degrees_y}
along the y dimension, about the geometry's own centroid. The geometry type
of the output matches the geometry type of the input.
}
\references{
\href{https://docs.rs/geo/latest/geo/algorithm/skew/trait.Skew.html}{Skew}
}
\seealso{
Other affine:
\code{\link[=rotate_around_center]{rotate_around_center()}},
\code{\link[=rotate_around_centroid]{rotate_around_centroid()}},
\code{\link[=scale_xy]{scale_xy()}},
\code{\link[=skew]{skew()}},
\code{\link[=translate]{translate()}}
}
\concept{affine}
